backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
greenlet>=3.0.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.17
httpx==0.28.1
aiosqlite==0.20.0
asyncpg>=0.30.0
pg8000>=1.31.0
redis>=5.0.0
boto3>=1.35.0
anyio>=4.0.0
